# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gig.PostGigRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      end_time:
        type: string
      pay_type:
        description: PER_DAY, PER_HOUR
        type: string
      skill_id:
        type: string
      start_time:
        type: string
      title:
        type: string
      wage_per_worker:
        type: integer
      workers_needed:
        type: integer
    required:
    - category_id
    - description
    - end_time
    - pay_type
    - skill_id
    - start_time
    - title
    - wage_per_worker
    - workers_needed
    type: object
  support.CreateTicketRequest:
    properties:
      description:
        type: string
      related_gig_id:
        type: string
      subject:
        type: string
    required:
    - description
    - subject
    type: object
...
  utils.SuccessResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@shiftley.com
    name: API Support
    url: http://www.shiftley.com/support
  description: Shiftley Backend API Documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Shiftley API
  version: "1.0"
paths:
...
  /applications/{applicationId}/status:
    patch:
      consumes:
      - application/json
      description: Hires a specific applicant for the gig slot.
      parameters:
      - description: Application ID
        in: path
        name: applicationId
        required: true
        type: string
      - description: 'Status: APPROVED'
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Approve Application
      tags:
      - Gig
...
  /gigs:
    post:
      consumes:
      - application/json
      description: Creates a gig and generates a mock Razorpay Escrow order.
      parameters:
      - description: Gig Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gig.PostGigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Post a New Gig
      tags:
      - Gig
  /gigs/{gigId}/applications:
    get:
      description: Fetches the list of employees who have applied for a specific gig.
      parameters:
      - description: Gig ID
        in: path
        name: gigId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: View Gig Applications
      tags:
      - Gig
  /gigs/{gigId}/attendance-qr:
    get:
      description: Generates a time-sensitive, rotating QR code string.
      parameters:
      - description: Gig ID
        in: path
        name: gigId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate Shift QR Codes
      tags:
      - Gig
  /gigs/{gigId}/cancel:
    post:
      consumes:
      - application/json
      description: Allows the employer to cancel a gig with automated penalty calculation.
      parameters:
      - description: Gig ID
        in: path
        name: gigId
        required: true
        type: string
      - description: Reason
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Cancel Gig
      tags:
      - Gig
  /gigs/{gigId}/close-unfilled:
    post:
      description: Closes a gig that had no worker approvals before start.
      parameters:
      - description: Gig ID
        in: path
        name: gigId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Close Gig with Full Refund
      tags:
      - Gig
  /gigs/{gigId}/emergency-hire:
    post:
      consumes:
      - application/json
      description: Penalizes the no-show employee and approves a replacement.
      parameters:
      - description: Gig ID
        in: path
        name: gigId
        required: true
        type: string
      - description: Replacement Details
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Emergency No-Show Replacement
      tags:
      - Gig
  /gigs/{gigId}/employees/{empId}/complete:
    post:
      consumes:
      - application/json
      description: Triggers the Razorpay payout for a worker.
      parameters:
      - description: Gig ID
        in: path
        name: gigId
        required: true
        type: string
      - description: Employee ID
        in: path
        name: empId
        required: true
        type: string
      - description: Override Details
        in: body
        name: request
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Manual Shift Complete (Trigger Payout)
      tags:
      - Gig
  /gigs/{gigId}/employees/{empId}/mark-arrived:
    post:
      description: Manually mark the worker as arrived via the app.
      parameters:
      - description: Gig ID
        in: path
        name: gigId
        required: true
        type: string
      - description: Employee ID
        in: path
        name: empId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Manual Check-In (QR Fallback)
      tags:
      - Gig
  /gigs/{gigId}/employees/{empId}/review:
    post:
      consumes:
      - application/json
      description: Allows the employer to leave a rating and comment for the employee.
      parameters:
      - description: Gig ID
        in: path
        name: gigId
        required: true
        type: string
      - description: Employee ID
        in: path
        name: empId
        required: true
        type: string
      - description: Review Details
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Rate Employee Post-Shift
      tags:
      - Gig
  /gigs/wage-benchmark:
    get:
      description: Returns the average market wage for a skill.
      parameters:
      - description: Skill ID
        in: query
        name: skill_id
        required: true
        type: string
      - description: PER_DAY or PER_HOUR
        in: query
        name: pay_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Wage Benchmark Check
      tags:
      - Gig
...
  /support/tickets:
    post:
      consumes:
      - application/json
      description: Creates a new support ticket for a user.
      parameters:
      - description: Ticket Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/support.CreateTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Raise Customer Support Ticket
      tags:
      - Support
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"